Done HW2.

+++ src/main/java/ru/javawebinar/topjava/repository/inmemory/InMemoryMealRepository.java
@@ -5,7 +5,7 @@ import ru.javawebinar.topjava.model.Meal;
 import ru.javawebinar.topjava.repository.MealRepository;
 import ru.javawebinar.topjava.util.MealsUtil;
 
-import java.util.Collection;
+import java.util.List;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -13,43 +13,49 @@ import java.util.stream.Collectors;
 
 @Repository
 public class InMemoryMealRepository implements MealRepository {
-    private Map<Integer, Meal> repository = new ConcurrentHashMap<>();
+    private final Map<Integer, Meal> repository = new ConcurrentHashMap<>();
     private final AtomicInteger counter = new AtomicInteger(0);
 
     {
-        repository=MealsUtil.meals.stream().collect(Collectors.toMap(Meal::getId, item -> item));
+        for (Meal meal : MealsUtil.meals) {
+            this.save(meal, meal.getUserId());
+        }
     }
 
     @Override
-    public Meal save(Meal meal,int userId) {
+    public Meal save(Meal meal, int userId) {
         if (meal.isNew()) {
             meal.setId(counter.incrementAndGet());
             meal.setUserId(userId);
             repository.put(meal.getId(), meal);
             return meal;
+        } else {
+            if (meal.getUserId().equals(userId)) {
+                return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+            }
         }
         // handle case: update, but not present in storage
-        return repository.computeIfPresent(meal.getId(), (id, oldMeal) -> meal);
+        return null;
     }
 
     @Override
-    public boolean delete(int id,int userId) {
-        Meal meal=repository.get(id);
-        if(meal.getUserId().equals(userId)){
+    public boolean delete(int id, int userId) {
+        Meal meal = repository.get(id);
+        if (meal != null && meal.getUserId().equals(userId)) {
             return repository.remove(id) != null;
         }
         return false;
     }
 
     @Override
-    public Meal get(int id,int userId) {
-        Meal meal=repository.get(id);
-        return (meal.getUserId().equals(userId))?meal:null;
+    public Meal get(int id, int userId) {
+        Meal meal = repository.get(id);
+        return (meal != null && meal.getUserId().equals(userId)) ? meal : null;
     }
 
     @Override
-    public Collection<Meal> getAll(int userId) {
-        return repository.values().stream().filter(x->x.getUserId().equals(userId)).sorted(((o1, o2) -> o2.getDateTime().compareTo(o1.getDateTime()))).collect(Collectors.toList());
+    public List<Meal> getAll(int userId) {
+        return repository.values().stream().filter(x -> x.getUserId().equals(userId)).sorted(((o1, o2) -> o2.getDateTime().compareTo(o1.getDateTime()))).collect(Collectors.toList());
     }
 }
